requirements.txt
defusedxml==0.5.0
dj-database-url==0.5.0
Django==1.10
django-bootstrap3==7.0.1
django-braces==1.9.0
django-cors-headers==1.3.1
django-oauth-toolkit==0.10.0
django-rest-framework-social-oauth2==1.0.4
djangorestframework==3.4.3
gunicorn==19.6.0
oauthlib==1.0.3
Pillow==5.1.0
PyJWT==1.4.2
python-social-auth==0.2.21
python3-openid==3.0.10
pytz==2016.6.1
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
stripe==1.37.0
whitenoise==3.2.1